\name{astercode}
\alias{astercode}
\title{Compile AsterCode
}
\description{Compile (source) the
  PEIP code in local session.  Used it the
  Package is not installed.
}
\usage{
astercode(dir = "RLIB")
}
\arguments{
  \item{dir}{directory where code is stored
}
}
\details{
  This function was created and used before the
  R-code was packaged as a library.
}
\value{side effects
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\keyword{misc}
