\name{bartl}
\alias{bartl}
\title{Bartlett window 
}
\description{Bartlett (triangle) window of length m
}
\usage{
bartl(m)
}
\arguments{
  \item{m}{integer, length of vector
}
}

\value{vector
}
\references{Aster, R.C., C.H. Thurber, and B. Borchers,
  \emph{Parameter Estimation and Inverse Problems}, Elsevier Academic Press, Amsterdam, 2005.
}
\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\examples{

bartl(11)

}
\keyword{misc}
