\name{GSVD}
\alias{GSVD}
\title{Generalized svd
}
\description{Generalized svd
}
\usage{
GSVD(A, B)
}
\arguments{
  \item{A}{U*E1*t(Q) }
  \item{B}{V*E2*t(Q)
}
}

\value{
   \item{U}{matrix}
  \item{V}{matrix}
  \item{X}{matrix}
  \item{C}{matrix}
  \item{S}{matrix}
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}
\note{lapack
}

\keyword{misc}
