\name{idcost}
\alias{idcost}
\title{Inverse  cosine transform 
}
\description{
  Takes the column-by-column inverse discrete cosine transform of Y.
}
\usage{
idcost(Y)
}
\arguments{
  \item{Y}{Input cosine transform
}
}

\value{Time series
}
\references{Aster, R.C., C.H. Thurber, and B. Borchers,
  \emph{Parameter Estimation and Inverse Problems}, Elsevier Academic Press, Amsterdam, 2005.
}
\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\seealso{dcost
}

\keyword{misc}
