\name{linesconst}
\alias{linesconst}
\title{Plot constant model
}
\description{
 Add to plotting  model in piecewise constant form over n subintervals,
 where n is the length of x.
 
}
\usage{
linesconst(x, l, r, ...)
}
\arguments{
  \item{x}{model to be plotted

}
  \item{l}{left endpoint of plot

}
  \item{r}{right endpoint of plot

}
  \item{\dots}{graphical parameters 
}
}
\details{Used for plotting vector models
}
\value{graphical side effects
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\seealso{plotconst
}

\keyword{misc}
