\name{tinv}
\alias{tinv}
\title{
  Inverse T-distribution
}
\description{Inverse T-distribution, qt
}
\usage{
tinv(p, nu)
}
\arguments{
  \item{p}{P-value
}
  \item{nu}{degrees of freedom
}
}
\details{Wrapper for qt
}
\value{Quantile for T-distribution
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\seealso{qt
}
\examples{

tinv(.4, 10)


}
\keyword{misc}
