\name{Vnorm}
\alias{Vnorm}
\title{Vector 2-Norm
}
\description{Vector 2-Norm.
}
\usage{
Vnorm(X)
}
\arguments{
  \item{X}{
    numeric vector
}
}

\value{Numeric scale norm
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}
\note{This function is intended to duplicated the
  matlab function norm.
}



\examples{

V = Vnorm(rnorm(10))


}
\keyword{misc}
