\name{imagesc}
\alias{imagesc}
\title{Image Display
}
\description{Display image in matlab format, i.e. flip and transpose.
}
\usage{
imagesc(G, col = grey((1:99)/100), ...)
}
\arguments{
  \item{G}{Image matrix
}
  \item{col}{color scale
}
  \item{\dots}{graphical parameters
}
}
\details{
  Program flips image and transposes.
}
\value{graphical side effects
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}

\keyword{misc}
